requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-celery==3.1.16
Django==1.7.4
djangorestframework==3.0.4
facebook-sdk==0.4.0
kombu==3.0.24
mysql-python==1.2.5
pytz==2014.10
requests==2.6.0
Unidecode==0.4.17
